A17 - Correção do tipo do atributo idade

+++ src/modulo02/aula17/PessoaFisica.java
@@ -5,7 +5,7 @@ public class PessoaFisica {
     private int codigo;
     private String nome;
     private String sobrenome;
-    private short idade;
+    private int idade;
     private String rg;
     private String cpf;
 
@@ -34,11 +34,11 @@ public class PessoaFisica {
         return sobrenome;
     }
 
-    public void setIdade(short idade) {
+    public void setIdade(int idade) {
         this.idade = idade;
     }
 
-    public short getIdade() {
+    public int getIdade() {
         return idade;
     }
